Use int32 for all test cases

+++ go/main.go
@@ -10,7 +10,7 @@ import (
 func main() {
 }
 
-type SortInt []int
+type SortInt []int32
 
 func (c SortInt) Len() int           { return len(c) }
 func (c SortInt) Swap(i, j int)      { c[i], c[j] = c[j], c[i] }
@@ -20,7 +20,7 @@ func (c SortInt) Less(i, j int) bool { return c[i] < c[j] }
 func testSort() {
 	arr := make(SortInt, 100_000)
 	for _i := range arr {
-		arr[_i] = rand.Intn(100)
+		arr[_i] = int32(rand.Intn(100))
 	}
 	temp := make(SortInt, len(arr))
 	for _i := 0; _i < 500; _i++ {
